Move shared WMR components into a shared assembly
