+++ Assets/MixedRealityToolkit.Providers/WindowsMixedReality/Shared.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cd8d49a94fcd2844da0a7a3d1de13252
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
