+++ Assets/MixedRealityToolkit.Providers/WindowsMixedReality/Shared/BaseWindowsMixedRealityCameraSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e643b8fb8328d7b498c993514cb7ca8f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
