+++ Assets/MixedRealityToolkit.Providers/WindowsMixedReality/Shared/Microsoft.MixedReality.Toolkit.Providers.WindowsMixedReality.Shared.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a37fab06e2e557e46930b8f918e51980
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
