+++ Assets/MixedRealityToolkit.Providers/WindowsMixedReality/Shared/WindowsMixedRealityCameraSettingsProfile.cs.meta
