+++ Assets/MixedRealityToolkit.Providers/WindowsMixedReality/Shared/WindowsMixedRealityHandRecorder.cs
@@ -2,17 +2,11 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using Microsoft.MixedReality.Toolkit.Utilities;
-using Microsoft.MixedReality.Toolkit.WindowsMixedReality.Input;
 using System;
-using System.Collections.Generic;
 using UnityEngine;
 
 #if WINDOWS_UWP
 using System.IO;
-using System.Threading.Tasks;
-using Windows.Storage;
-using Windows.Storage.Pickers;
-using Windows.Storage.Provider;
 #endif
 
 namespace Microsoft.MixedReality.Toolkit.Input
@@ -55,7 +49,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
 
         private void RecordHandStart(Handedness handedness)
         {
-            HandJointUtils.TryGetJointPose<WindowsMixedRealityArticulatedHand>(ReferenceJoint, handedness, out MixedRealityPose joint);
+            HandJointUtils.TryGetJointPose(ReferenceJoint, handedness, out MixedRealityPose joint);
             offset = joint.Position;
             recordingHand = handedness;
         }
@@ -65,7 +59,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
             MixedRealityPose[] jointPoses = new MixedRealityPose[jointCount];
             for (int i = 0; i < jointCount; ++i)
             {
-                HandJointUtils.TryGetJointPose<WindowsMixedRealityArticulatedHand>((TrackedHandJoint)i, recordingHand, out jointPoses[i]);
+                HandJointUtils.TryGetJointPose((TrackedHandJoint)i, recordingHand, out jointPoses[i]);
             }
 
             ArticulatedHandPose pose = new ArticulatedHandPose();
@@ -77,7 +71,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
             StoreRecordedHandPose(pose.ToJson(), filename);
         }
 
-        #if WINDOWS_UWP
+#if WINDOWS_UWP
         private static void StoreRecordedHandPose(string data, string filename)
         {
             string path = Path.Combine(Application.persistentDataPath, filename);
@@ -86,12 +80,12 @@ namespace Microsoft.MixedReality.Toolkit.Input
                 writer.Write(data);
             }
         }
-        #else
+#else
         private static void StoreRecordedHandPose(string data, string filename)
         {
             Debug.Log(data);
         }
-        #endif
+#endif
     }
 
 }
