+++ Assets/MixedRealityToolkit.Providers/WindowsMixedReality/Shared/WindowsMixedRealityReprojectionUpdater.cs
